Update/dependencies

+++ core/src/database/mongo.ts
@@ -3,9 +3,11 @@ import * as mongo from 'mongodb';
 import { IProvider, IRecoveryCode, IUser, IUserUpdate } from '../types';
 import { IDatabaseAdapter } from './adapter';
 
+const UNINITIALIZED_ADAPTER = 'Mongo adapter was not initialized';
+
 export default class MongoAdapter implements IDatabaseAdapter {
   private databaseURL: string
-  private db: mongo.Db
+  private db: mongo.Db | undefined
 
   constructor(databaseURL: string) {
     this.databaseURL = databaseURL
@@ -22,16 +24,19 @@ export default class MongoAdapter implements IDatabaseAdapter {
   }
 
   public async findUserByEmail(email: string): Promise<IUser | undefined> {
+    if(!this.db) throw new Error(UNINITIALIZED_ADAPTER);
     return this.normalize(await this.db.collection('auth_users').findOne({ email }))
   }
 
   public async findUserById(id: string): Promise<IUser | undefined> {
+    if(!this.db) throw new Error(UNINITIALIZED_ADAPTER);
     return this.normalize(
       await this.db.collection('auth_users').findOne({ _id: new mongo.ObjectID(id) })
     )
   }
 
   public async findUserByProviderLogin(login: string): Promise<IUser | undefined> {
+    if(!this.db) throw new Error(UNINITIALIZED_ADAPTER);
     const provider = await this.db.collection('auth_providers').findOne({ login })
     if (!provider) {
       return undefined
@@ -42,6 +47,7 @@ export default class MongoAdapter implements IDatabaseAdapter {
   }
 
   public findRecoveryCodesByUserId(userId: string): Promise<IRecoveryCode[]> {
+    if(!this.db) throw new Error(UNINITIALIZED_ADAPTER);
     return this.db
       .collection('auth_recovery_codes')
       .find({ userId })
@@ -49,10 +55,12 @@ export default class MongoAdapter implements IDatabaseAdapter {
   }
 
   public async insertRecoveryCodes(userId: string, codes: string[]): Promise<IRecoveryCode[]> {
+    if(!this.db) throw new Error(UNINITIALIZED_ADAPTER);
     await this.db.collection('auth_recovery_codes').deleteMany({ userId })
 
     await Promise.all(
       codes.map(code => {
+        if(!this.db) throw new Error(UNINITIALIZED_ADAPTER);
         return this.db.collection('auth_recovery_codes').insertOne({ userId, code, used: false })
       })
     )
@@ -60,6 +68,7 @@ export default class MongoAdapter implements IDatabaseAdapter {
   }
 
   public async useRecoveryCode(userId: string, code: string): Promise<boolean> {
+    if(!this.db) throw new Error(UNINITIALIZED_ADAPTER);
     const res = await this.db
       .collection('auth_recovery_codes')
       .updateOne({ userId, code: code.toLowerCase(), used: false }, { $set: { used: true } })
@@ -67,11 +76,13 @@ export default class MongoAdapter implements IDatabaseAdapter {
   }
 
   public async insertUser(user: IUser): Promise<string> {
+    if(!this.db) throw new Error(UNINITIALIZED_ADAPTER);
     const res = await this.db.collection('auth_users').insertOne(user)
     return res.insertedId.toHexString()
   }
 
   public async updateUser(user: IUserUpdate): Promise<void> {
+    if(!this.db) throw new Error(UNINITIALIZED_ADAPTER);
     const res = await this.db
       .collection('auth_users')
       .update({ _id: new mongo.ObjectID(user.id!) }, { $set: omit(user, 'id') })
@@ -79,6 +90,7 @@ export default class MongoAdapter implements IDatabaseAdapter {
   }
 
   public async insertProvider(provider: IProvider): Promise<void> {
+    if(!this.db) throw new Error(UNINITIALIZED_ADAPTER);
     await this.db.collection('auth_providers').insertOne(provider)
     // return res.insertedId.toHexString()
   }
